Make README prettier by including names, not links
I found the list a bit hard to read as each entry was just a GitHub link being dropped in for each point with many `https://github.com/`  repeated.

I think now the list looks great. And contains the names of the authors of these awesome repositories.

Also as I was going through the list, I removed `https://github.com/sotojuan` as the repo no longer exists and the author stopped sharing it.

I hope you consider adding this change. Hope more people add their names to the list in the future too. 🙂